requirements.txt
aiobotocore==3.1.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
asttokens==3.0.1
attrs==25.4.0
botocore==1.42.42
certifi==2026.1.4
charset-normalizer==3.4.4
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
duckdb==1.4.4
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2026.2.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.1
nbclient==0.10.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==3.0.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3fs==2026.2.0
setuptools==70.2.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch
torchaudio
torchvision
tornado
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
uv==0.10.0
wcwidth==0.6.0
wrapt==2.1.1
yarl==1.22.0
pyyaml
torchmetrics
faster-coco-eval